Migrate to github

+++ MMALSharp/Native/MMALUtil.cs
@@ -16,11 +16,11 @@ namespace MMALSharp.Native
 
         public static uint MMAL_FOURCC(string s)
         {
-            int a1 = s[0];
-            int b1 = s[1];
-            int c1 = s[2];
-            int d1 = s[3];
-            return (uint) (((a1) | (b1 << 8) | (c1 << 16) | (d1 << 24)));
+            uint a1 = s[0];
+            uint b1 = s[1];
+            uint c1 = s[2];
+            uint d1 = s[3];
+            return ((a1) | (b1 << 8) | (c1 << 16) | (d1 << 24));
         }
 
         public enum MMAL_STATUS_T
